Add separation of concerns
